# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TutorTurnRequest": {
        "properties": {
          "client_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Revision Id"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "enum": [
              "test",
              "practice"
            ],
            "title": "Mode",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "step_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Index"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "TutorTurnRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend for the Quantum Learning Laboratory.\n\nThe quantum engine computes, the authored evaluator grades, and the tutor\nteaches. Private rubrics and answer keys are never returned by any route in\nthis contract.",
    "title": "Quantum Learning Laboratory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tutor/conversations/{topic_key}": {
      "get": {
        "description": "Owner-scoped history. A guessed conversation key returns nothing else's data.",
        "operationId": "read_conversation_api_v1_tutor_conversations__topic_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_key",
            "required": true,
            "schema": {
              "title": "Topic Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Conversation Api V1 Tutor Conversations  Topic Key  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Conversation",
        "tags": [
          "tutor"
        ]
      }
    },
...
    "/api/v1/tutor/hint-availability/{topic_id}": {
      "get": {
        "operationId": "hint_availability_api_v1_tutor_hint_availability__topic_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Hint Availability Api V1 Tutor Hint Availability  Topic Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hint Availability",
        "tags": [
          "tutor"
        ]
      }
    },
    "/api/v1/tutor/passages/{passage_id}": {
      "get": {
        "description": "Open a cited passage without discarding the learner's work.",
        "operationId": "read_passage_api_v1_tutor_passages__passage_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "passage_id",
            "required": true,
            "schema": {
              "title": "Passage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Passage Api V1 Tutor Passages  Passage Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Passage",
        "tags": [
          "tutor"
        ]
      }
    },
    "/api/v1/tutor/turns": {
      "post": {
        "description": "One tutor turn. Scope is re-decided here on every request.",
        "operationId": "create_turn_api_v1_tutor_turns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Turn Api V1 Tutor Turns Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Turn",
        "tags": [
          "tutor"
        ]
      }
    }
  }
}